pyproject.toml
[tool.poetry]
name = "de-hw2"
version = "0.1.0"
description = ""
authors = ["Dmitry <kachkindmitry@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
medical-data-processor = "med_results_parser.main:main"

[tool.poetry.dependencies]
python = "^3.10"
polars = "^1.17.1"
psycopg2-binary = "^2.9.10"
pydantic = "^2.10.4"
pydantic-settings = "^2.7.0"
fastexcel = "^0.12.0"
pyaml = "^24.12.1"
xlsxwriter = "^3.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.11.0"
pre-commit = "^3.5.0"
pylint = "^3.0.2"
isort = "^5.12.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3


[tool.mypy]
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
warn_unreachable = true
pretty = true

[[tool.mypy.overrides]]
module = "*.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.pylint.main]
fail-under = 10
ignore = ["data"]
ignored-modules = ["test"]
py-version = "3.10"
suggestion-mode = true

[tool.pylint.basic]
bad-names = ["foo", "bar", "baz", "toto", "tutu", "tata"]
class-attribute-naming-style = "any"
class-const-naming-style = "UPPER_CASE"
class-naming-style = "PascalCase"
const-naming-style = "UPPER_CASE"
function-naming-style = "snake_case"
property-classes = ["abc.abstractproperty"]


[tool.pylint.format]
ignore-long-lines = "^\\s*(# )?<?https?://\\S+>?$"
indent-after-paren = 4
indent-string = "    "
max-line-length = 79
max-module-lines = 1000


[tool.pylint.messages_control]
disable = [
    "missing-docstring",
    "too-few-public-methods",
    "too-many-arguments",
    "invalid-all-format",
    "import-error",
    "logging-fstring-interpolation"
]


good-names = []



[tool.pylint.similarities]
ignore-imports = true
ignore-docstrings = true
ignore-comments = true


[tool.pytest.ini_options]
asyncio_mode = "strict"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(levelname)-5.5s [%(name)s] %(message)s"




